feat: 소셜링크 등록, 해제

+++ src/main/java/com/YYduo/KkuldongVarietyStore/domain/member/dto/MemberResponseDto.java
@@ -20,7 +20,7 @@ public class MemberResponseDto {
 
     private String motto;
 
-    private Long avatar;
+    private int avatar;
 
     private String instagram;
